fix windows compilation

+++ core/src/string_utils.cpp
@@ -75,7 +75,7 @@ namespace Drill
                 char length = sizeof(suffix) / sizeof(suffix[0]);
 
                 int i = 0;
-                double dblBytes = bytes;
+                double dblBytes = (double)bytes;
 
                 if (bytes > 1024) {
                     for (i = 0; (bytes / 1024) > 0 && i<length-1; i++, bytes /= 1024)
